Refactor /cleanup
