+++ matico_components/src/Components/MaticoEditor/Panes/MaticoOutlineViewer/MaticoOutlineViewer.tsx
@@ -1,0 +1,135 @@
+import React, { useCallback, useState } from "react";
+import { Flex, Heading, View } from "@adobe/react-spectrum";
+import { useApp } from "Hooks/useApp";
+import { Page, PaneRef } from "@maticoapp/matico_types/spec";
+import { withRouter, RouteComponentProps } from "react-router-dom";
+import {
+    CollisionDetection,
+    DndContext,
+    DragEndEvent,
+    DragOverEvent,
+    DragOverlay,
+    KeyboardSensor,
+    MouseSensor,
+    TouchSensor,
+    useSensor,
+    useSensors
+} from "@dnd-kit/core";
+import { restrictToVerticalAxis } from "@dnd-kit/modifiers";
+import { coordinateGetter } from "../../EditorComponents/SortableDraggableList/multipleContainersKeyboardCoordinates";
+import throttle from "lodash/throttle";
+import { createPortal } from "react-dom";
+import { outlineCollisionDetection } from "./CollisionDetection";
+import { DraggingProvider } from "./DraggingContext";
+import { DraggablePane } from "./DraggablePane";
+import { PageList } from "./PageList";
+
+type MaticoOutlineViewerProps = RouteComponentProps & {};
+
+export const MaticoOutlineViewer: React.FC = withRouter(
+    ({ history, location }: MaticoOutlineViewerProps) => {
+        // list pages
+        const { pages, reparentPane, changePaneIndex } = useApp();
+        const [activeItem, setActiveItem] =
+            useState<Page | PaneRef | null>(null);
+        const handleDrag = throttle(
+            (event: DragOverEvent | DragEndEvent, isDragEnd: boolean) => {
+                const { over, active } = event;
+                const isSelf = active.id === over?.id;
+                if (isSelf || !over) return;
+                const currentParent = active?.data?.current?.parent;
+                const overParent = over?.data?.current?.parent;
+                const overNewParent =
+                    (over?.data?.current?.type === "page" ||
+                        over?.data?.current?.type === "container") &&
+                    currentParent?.id !== over?.id;
+                const haveParents = currentParent && overParent;
+                const overCousinRow =
+                    haveParents && currentParent.id !== overParent.id;
+                const isSibling =
+                    haveParents && currentParent.id === overParent.id;
+
+                if (overNewParent) {
+                    const paneRefId = active?.data?.current?.paneId;
+                    const targetId = over?.id as string;
+                    if (isDragEnd || over?.data?.current?.depth === 0) {
+                        reparentPane(paneRefId, targetId);
+                    }
+                    return;
+                } else if (overCousinRow) {
+                    const paneRefId = active?.data?.current?.paneId;
+                    const targetId = overParent?.id;
+                    reparentPane(paneRefId, targetId);
+                    return;
+                } else if (isSibling && isDragEnd) {
+                    const newIndex = over?.data?.current?.index;
+                    if (
+                        active?.data?.current?.paneId &&
+                        newIndex !== undefined
+                    ) {
+                        changePaneIndex(
+                            active?.data?.current?.paneId,
+                            newIndex
+                        );
+                    }
+                }
+            },
+            125
+        );
+
+        //@ts-ignore
+        const handleDragStart = ({ active }) => setActiveItem(active);
+        const handleDragOver = (event: DragOverEvent) =>
+            handleDrag(event, false);
+        const handleDragEnd = (event: DragEndEvent) => {
+            handleDrag(event, true);
+            setActiveItem(null);
+        };
+
+        const collisionDetectionStrategy: CollisionDetection = useCallback(
+            outlineCollisionDetection,
+            [activeItem, JSON.stringify(pages)]
+        );
+
+        const sensors = useSensors(
+            useSensor(MouseSensor),
+            useSensor(TouchSensor),
+            useSensor(KeyboardSensor, {
+                coordinateGetter
+            })
+        );
+
+        return (
+            <DraggingProvider activeItem={activeItem}>
+                <View width="100%" height="auto">
+                    <Flex direction="column">
+                        <Heading margin="size-150" marginBottom="size-0" alignSelf="start">
+                            Page Outline
+                        </Heading>
+                        <DndContext
+                            modifiers={[restrictToVerticalAxis]}
+                            onDragStart={handleDragStart}
+                            onDragEnd={handleDragEnd}
+                            // onDragOver={handleDragOver}
+                            collisionDetection={collisionDetectionStrategy}
+                            sensors={sensors}
+                        >
+                            {pages.map((page) => (
+                                <PageList key={page.id} page={page} />
+                            ))}
+                            {!!activeItem &&
+                                createPortal(
+                                    <DragOverlay adjustScale={false}>
+                                        <DraggablePane
+                                            activeItem={activeItem}
+                                        />
+                                    </DragOverlay>,
+                                    document.body
+                                )}
+                        </DndContext>
+                    </Flex>
+                </View>
+            </DraggingProvider>
+        );
+    }
+);
