+++ matico_components/src/Components/MaticoEditor/Panes/MaticoOutlineViewer/index.tsx
@@ -1,0 +1,5 @@
+import { MaticoOutlineViewer } from "./MaticoOutlineViewer";
+
+export {
+    MaticoOutlineViewer
+}
